refactor(components)
* add jsconfig.json file
* update .eslintrc file to resolve imports in components
* simplify imports in components
