+++ src/components/Menu/Menu.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import { routes } from '../../routes/routes';
+import { routes } from 'routes';
 
 import './menu.css';
