+++ src/index.jsx
@@ -6,7 +6,7 @@ import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import { App } from './App';
 import * as serviceWorker from './serviceWorker';
-import { rootReducer } from './redux/reducers/rootReducer';
+import { rootReducer } from './store/reducers/rootReducer';
 
 const store = createStore(
   rootReducer,
