+++ src/routes/Router.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import { routes } from './routes';
+import { routes } from 'routes';
 
 export const Router = () => (
   <Switch>
