+++ src/routes/index.js
@@ -1,0 +1,2 @@
+export { routes } from 'routes/routes';
+export { Router } from 'routes/Router';
